Revert "Webhook class is implemented"
This reverts commit 91e8ead0522667abe7d95e9065283cf8e1f1dd93.
